+++ plone/app/content/browser/file.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
+from AccessControl import getSecurityManager
 from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.permissions import AddPortalContent
 from Products.Five.browser import BrowserView
 from plone.app.dexterity.interfaces import IDXFileFactory
 from plone.dexterity.interfaces import IDexterityFTI
@@ -89,6 +91,13 @@ class FileUploadView(BrowserView):
             raise KeyError
 
     def __call__(self):
+        # Check if user has permission to add content here
+        sm = getSecurityManager()
+        if not sm.checkPermission(AddPortalContent, self.context):
+            response = self.request.RESPONSE
+            response.setStatus(403)
+            return "You are not authorized to add content to this folder."
+
         req = self.request
         tusrequest = False
         if TUS_ENABLED:
@@ -122,6 +131,17 @@ class FileUploadView(BrowserView):
         # Determine if the default file/image types are DX or AT based
         ctr = getToolByName(self.context, 'content_type_registry')
         type_ = ctr.findTypeName(filename.lower(), '', '') or 'File'
+
+        # Now check that the object is not restricted to be added in the
+        # current context
+        allowed_ids = [fti.getId() for fti in self.context.allowedContentTypes()]
+        if type_ not in allowed_ids:
+            response = self.request.RESPONSE
+            response.setStatus(403)
+            if type_ == 'File':
+                return "You cannot add a File to this folder, try another one"
+            if type_ == 'Image':
+                return "You cannot add an Image to this folder, try another one"
 
         dx_based = False
         pt = getToolByName(self.context, 'portal_types')